scripts/requirements.txt
pandas==0.25.2
numpy==1.20.0
selenium==4.1.3
webdriver-manager==3.5.4
pycountry==22.3.5
beautifulsoup4==4.11.1
lxml==4.8.0
PyYAML==5.1.2
Flask==2.1.2
Flask-RESTful==0.3.9
pyspark==2.4.4
pyarrow==0.16.0